Implement storage abstraction layer with trait-based design
• Add TopicLogStorage trait for pluggable storage backends
• Refactor TopicLog to InMemoryTopicLog implementation
• Update FlashQ to use Box<dyn TopicLogStorage> abstraction
• Add StorageBackend enum for future storage selection
