+++ src/lib.rs
@@ -2,10 +2,12 @@ use std::collections::HashMap;
 use std::collections::hash_map::Entry::{Occupied, Vacant};
 use std::fmt;
 use std::sync::{Arc, Mutex};
+use storage::TopicLog;
 
 pub mod demo;
 #[cfg(feature = "http")]
 pub mod http;
+pub mod storage;
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum FlashQError {
@@ -109,64 +111,6 @@ impl RecordWithOffset {
 // =============================================================================
 
 #[derive(Debug, Clone)]
-pub struct TopicLog {
-    records: Vec<RecordWithOffset>,
-    next_offset: u64,
-}
-
-impl Default for TopicLog {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl TopicLog {
-    pub fn new() -> Self {
-        TopicLog {
-            records: Vec::new(),
-            next_offset: 0,
-        }
-    }
-
-    pub fn append(&mut self, record: Record) -> u64 {
-        let current_offset = self.next_offset;
-        let record_with_offset = RecordWithOffset::from_record(record, current_offset);
-        self.records.push(record_with_offset);
-        self.next_offset += 1;
-        current_offset
-    }
-
-    pub fn get_records_from_offset(
-        &self,
-        offset: u64,
-        count: Option<usize>,
-    ) -> Vec<&RecordWithOffset> {
-        let start_index = offset as usize;
-        if start_index >= self.records.len() {
-            return Vec::new();
-        }
-        let slice = &self.records[start_index..];
-        let limited_slice = match count {
-            Some(limit) => &slice[..limit.min(slice.len())],
-            None => slice,
-        };
-        limited_slice.iter().collect()
-    }
-
-    pub fn len(&self) -> usize {
-        self.records.len()
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.records.is_empty()
-    }
-
-    pub fn next_offset(&self) -> u64 {
-        self.next_offset
-    }
-}
-
-#[derive(Debug, Clone)]
 pub struct ConsumerGroup {
     group_id: String,
     topic_offsets: HashMap<String, u64>,
@@ -194,7 +138,7 @@ impl ConsumerGroup {
 }
 
 pub struct FlashQ {
-    topics: Arc<Mutex<HashMap<String, TopicLog>>>,
+    topics: Arc<Mutex<HashMap<String, Box<dyn TopicLog>>>>,
     consumer_groups: Arc<Mutex<HashMap<String, ConsumerGroup>>>,
 }
 
@@ -214,13 +158,17 @@ impl FlashQ {
 
     pub fn post_record(&self, topic: String, record: Record) -> Result<u64, String> {
         let mut topic_log_map = self.topics.lock().unwrap();
-        let topic_log = topic_log_map.entry(topic).or_default();
+        let topic_log = topic_log_map
+            .entry(topic)
+            .or_insert_with(|| storage::StorageBackend::Memory.create());
         Ok(topic_log.append(record))
     }
 
     pub fn post_records(&self, topic: String, records: Vec<Record>) -> Result<Vec<u64>, String> {
         let mut topic_log_map = self.topics.lock().unwrap();
-        let topic_log = topic_log_map.entry(topic).or_default();
+        let topic_log = topic_log_map
+            .entry(topic)
+            .or_insert_with(|| storage::StorageBackend::Memory.create());
 
         let mut offsets = Vec::new();
         for record in records {
@@ -556,123 +504,6 @@ mod tests {
         assert_eq!(msg.record.value, "record with metadata");
         assert_eq!(msg.record.headers.as_ref().unwrap(), &headers);
         assert_eq!(msg.offset, 0);
-    }
-
-    // =============================================================================
-    // TOPIC LOG TESTS
-    // =============================================================================
-
-    // TopicLog Tests
-    #[test]
-    fn test_topic_log_creation() {
-        let log = TopicLog::new();
-        assert_eq!(log.len(), 0);
-        assert_eq!(log.next_offset(), 0);
-    }
-
-    #[test]
-    fn test_topic_log_append_single_record() {
-        let mut log = TopicLog::new();
-        let record = Record::new(None, "first record".to_string(), None);
-        let offset = log.append(record);
-
-        assert_eq!(offset, 0);
-        assert_eq!(log.len(), 1);
-        assert_eq!(log.next_offset(), 1);
-    }
-
-    #[test]
-    fn test_topic_log_append_multiple_records() {
-        let mut log = TopicLog::new();
-        let record1 = Record::new(None, "record 1".to_string(), None);
-        let record2 = Record::new(None, "record 2".to_string(), None);
-        let record3 = Record::new(None, "record 3".to_string(), None);
-
-        let offset1 = log.append(record1);
-        let offset2 = log.append(record2);
-        let offset3 = log.append(record3);
-
-        assert_eq!(offset1, 0);
-        assert_eq!(offset2, 1);
-        assert_eq!(offset3, 2);
-        assert_eq!(log.len(), 3);
-        assert_eq!(log.next_offset(), 3);
-    }
-
-    #[test]
-    fn test_topic_log_get_records_from_beginning() {
-        let mut log = TopicLog::new();
-        let record1 = Record::new(None, "first".to_string(), None);
-        let record2 = Record::new(None, "second".to_string(), None);
-        let record3 = Record::new(None, "third".to_string(), None);
-
-        log.append(record1);
-        log.append(record2);
-        log.append(record3);
-
-        let records = log.get_records_from_offset(0, None);
-        assert_eq!(records.len(), 3);
-        assert_eq!(records[0].record.value, "first");
-        assert_eq!(records[1].record.value, "second");
-        assert_eq!(records[2].record.value, "third");
-    }
-
-    #[test]
-    fn test_topic_log_get_records_from_middle_offset() {
-        let mut log = TopicLog::new();
-        let record1 = Record::new(None, "first".to_string(), None);
-        let record2 = Record::new(None, "second".to_string(), None);
-        let record3 = Record::new(None, "third".to_string(), None);
-
-        log.append(record1);
-        log.append(record2);
-        log.append(record3);
-
-        let records = log.get_records_from_offset(1, None);
-        assert_eq!(records.len(), 2);
-        assert_eq!(records[0].record.value, "second");
-        assert_eq!(records[1].record.value, "third");
-    }
-
-    #[test]
-    fn test_topic_log_get_records_with_count_limit() {
-        let mut log = TopicLog::new();
-        let record1 = Record::new(None, "first".to_string(), None);
-        let record2 = Record::new(None, "second".to_string(), None);
-        let record3 = Record::new(None, "third".to_string(), None);
-
-        log.append(record1);
-        log.append(record2);
-        log.append(record3);
-
-        let records = log.get_records_from_offset(0, Some(2));
-        assert_eq!(records.len(), 2);
-        assert_eq!(records[0].record.value, "first");
-        assert_eq!(records[1].record.value, "second");
-    }
-
-    #[test]
-    fn test_topic_log_get_records_beyond_log() {
-        let mut log = TopicLog::new();
-        let record = Record::new(None, "only record".to_string(), None);
-        log.append(record);
-
-        let records = log.get_records_from_offset(5, None);
-        assert_eq!(records.len(), 0);
-    }
-
-    #[test]
-    fn test_topic_log_record_offsets_match() {
-        let mut log = TopicLog::new();
-        let record1 = Record::new(None, "msg1".to_string(), None);
-        let record2 = Record::new(None, "msg2".to_string(), None);
-
-        let offset1 = log.append(record1);
-        let offset2 = log.append(record2);
-
-        let records = log.get_records_from_offset(0, None);
-        assert_eq!(records[0].offset, offset1);
-        assert_eq!(records[1].offset, offset2);
     }
 
     // =============================================================================
